+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.example.spinerproject">
+          package="com.example.todolist">
 
     <application
             android:allowBackup="true"
@@ -10,11 +10,11 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
 
-        <activity android:name="com.example.spinerproject.Settings">
+        <activity android:name="com.example.todolist.Settings">
         </activity>
-        <activity android:name="com.example.spinerproject.NewTask">
+        <activity android:name="com.example.todolist.NewTask">
         </activity>
-        <activity android:name="com.example.spinerproject.MainActivity">
+        <activity android:name="com.example.todolist.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
