+++ app/src/main/java/com/example/todolist/Settings.kt
@@ -1,4 +1,4 @@
-package com.example.spinerproject
+package com.example.todolist
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
