+++ app/src/main/java/com/example/todolist/model/Task.kt
@@ -1,4 +1,4 @@
-package model
+package com.example.todolist.model
 
 import java.util.*
