+++ app/src/main/res/layout/activity_new_task.xml
@@ -7,72 +7,44 @@
         android:layout_height="match_parent"
         android:background="#ffffff"
         android:orientation="vertical"
-        tools:context=".NewTask" android:onClick="findTask" android:id="@+id/addTask">
+        tools:context="com.example.todolist.NewTask" android:id="@+id/addTask">
 
 
-    <EditText
-            android:layout_width="352dp"
-            android:layout_height="80dp"
-            android:inputType="textPersonName"
-            android:ems="10"
-            android:id="@+id/titleNewTask" android:hint="@string/tasktitle" android:textColor="#090808"
-            android:textColorHint="#0C0B0B" android:textSize="18sp"
-            android:layout_marginEnd="8dp" app:layout_constraintEnd_toEndOf="parent" android:layout_marginStart="8dp"
-            app:layout_constraintStart_toStartOf="parent" android:layout_marginBottom="8dp"
-            app:layout_constraintBottom_toTopOf="@+id/descripNewTask" android:layout_marginTop="8dp"
-            app:layout_constraintTop_toBottomOf="@+id/tv1" app:layout_constraintHorizontal_bias="0.488"
-            app:layout_constraintVertical_bias="0.343"/>
-    <EditText
-            android:layout_width="352dp"
-            android:layout_height="80dp"
-            android:inputType="textMultiLine"
-            android:gravity="start|top"
-            android:ems="10"
-            android:id="@+id/descripNewTask" android:hint="@string/descNewTask" android:textColorHint="#070707"
-            android:textSize="18sp" android:textColor="#090808"
-            android:layout_marginEnd="8dp"
-            app:layout_constraintEnd_toEndOf="parent" android:layout_marginStart="8dp"
-            app:layout_constraintStart_toStartOf="parent"
-            android:layout_marginBottom="384dp"
-            app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintHorizontal_bias="0.581"/>
-    <TextView
-            android:text="@string/newTask"
-            android:layout_width="345dp"
-            android:layout_height="60dp" android:id="@+id/tv1" android:textColor="#090808" android:textSize="18sp"
-            android:layout_marginEnd="8dp" app:layout_constraintEnd_toEndOf="parent" android:layout_marginStart="8dp"
-            app:layout_constraintStart_toStartOf="parent"
-            android:visibility="visible" android:layout_marginTop="44dp" app:layout_constraintTop_toTopOf="parent"
-            android:layout_marginBottom="8dp" app:layout_constraintBottom_toTopOf="@+id/titleNewTask"/>
-
-    <Button
-            android:text="@string/addBT"
-            android:layout_width="199dp"
-            android:layout_height="53dp" android:id="@+id/addBT"
-            android:layout_marginEnd="8dp" app:layout_constraintEnd_toEndOf="parent" android:layout_marginStart="8dp"
-            app:layout_constraintStart_toStartOf="parent" android:layout_marginBottom="16dp"
-            app:layout_constraintBottom_toTopOf="@+id/BtFindTask" app:layout_constraintHorizontal_bias="0.397"
-            android:onClick="addDatos"/>
-    <Button
-            android:text="@string/findTask"
-            android:layout_width="199dp"
-            android:layout_height="53dp" android:id="@+id/BtFindTask" android:layout_alignParentEnd="true"
-            android:layout_alignBottom="@+id/addBT"
-            android:layout_marginEnd="8dp" app:layout_constraintEnd_toEndOf="parent" android:layout_marginStart="8dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintHorizontal_bias="0.397"
-            android:onClick="findTask" android:layout_marginTop="128dp"
-            app:layout_constraintTop_toBottomOf="@+id/descripNewTask" android:layout_marginBottom="16dp"
-            app:layout_constraintBottom_toTopOf="@+id/button3" app:layout_constraintVertical_bias="1.0"/>
-    <Button
-            android:text="@string/addBack"
-            android:layout_width="199dp"
-            android:layout_height="53dp" android:id="@+id/button3" android:layout_alignParentEnd="true"
-            android:layout_alignParentBottom="true"
-            android:layout_marginEnd="8dp"
-            app:layout_constraintEnd_toEndOf="parent" android:layout_marginStart="8dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintHorizontal_bias="0.397" android:layout_marginBottom="88dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            android:onClick="back"/>
+    <LinearLayout
+            android:orientation="vertical"
+            android:layout_width="335dp"
+            android:layout_height="426dp" android:layout_marginBottom="8dp"
+            app:layout_constraintBottom_toBottomOf="parent" android:layout_marginTop="8dp"
+            app:layout_constraintTop_toTopOf="parent" app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginEnd="8dp" app:layout_constraintStart_toStartOf="parent"
+            android:layout_marginStart="8dp" android:id="@+id/linearLayout">
+        <TextView
+                android:text="@string/newTask"
+                android:layout_width="314dp"
+                android:layout_height="54dp" android:id="@+id/tvIntroduceDates" android:textColor="#090808"
+                android:textSize="18sp"
+                android:layout_marginEnd="8dp"
+                android:visibility="visible"
+                android:layout_marginStart="8dp" android:textAlignment="center"/>
+        <EditText
+                android:layout_width="match_parent"
+                android:layout_height="90sp"
+                android:inputType="textPersonName"
+                android:ems="10"
+                android:id="@+id/titleNewTask" android:hint="@string/tasktitle" android:textColor="#090808"
+                android:textColorHint="#0C0B0B" android:textSize="18sp"
+                android:layout_marginEnd="8dp" android:layout_marginStart="8dp"
+        />
+        <EditText
+                android:layout_width="match_parent"
+                android:layout_height="90sp"
+                android:inputType="textMultiLine"
+                android:gravity="start|top"
+                android:ems="10"
+                android:id="@+id/descripNewTask" android:hint="@string/descNewTask" android:textColorHint="#070707"
+                android:textSize="18sp" android:textColor="#090808"
+                android:layout_marginEnd="8dp"
+                android:layout_marginStart="8dp"/>
+    </LinearLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
