+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".Settings" android:background="#ffffff">
+        tools:context="com.example.todolist.Settings" android:background="#ffffff">
 
     <TextView
             android:text="@string/setting"
