+++ app/src/main/res/layout/list_item_tareas.xml
@@ -16,7 +16,7 @@
         <TableRow android:layout_width="match_parent" android:layout_height="match_parent">
             <CheckBox
                     android:layout_width="match_parent"
-                    android:layout_height="wrap_content" android:id="@+id/check1" android:onClick="deleteTask"
+                    android:layout_height="wrap_content" android:id="@+id/check1" android:onClick="deletedTask"
                     android:background="#ffffff" android:textColor="#ffffff" android:textColorHint="#ffffff"
                     android:drawableTint="#ffffff" android:focusable="false" android:focusableInTouchMode="false"/>
             <TextView
