+++ app/src/main/res/menu/addmenu.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:app="http://schemas.android.com/apk/res-auto" xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:title="Item" android:icon="?attr/actionModeCloseDrawable" app:showAsAction="always"
+          android:id="@+id/back"/>
+    <item android:title="Item" android:icon="@android:drawable/ic_menu_save" app:showAsAction="always"
+          android:id="@+id/save"/>
+</menu>
