+++ app/src/main/res/values/strings.xml
@@ -12,7 +12,7 @@
     <string name="descNewTask">Description</string>
     <string name="dateTask">Date</string>
     <string name="timeTask">Time</string>
-    <string name="newTask">New Task</string>
+    <string name="newTask">Introduce new task dates</string>
     <string name="findTask">Find Task</string>
     <string name="setting">Settings</string>
     <string name="tvMyTask">My Tasks</string>
